Move `CheckoutMessage` to checkout subfolder
